refactor dpm monitoring routes
