+++ src/app/routers/dpm_command_center.py
@@ -1,11 +1,7 @@
 from fastapi import APIRouter, Path, Query
 
 from app.contracts.dpm_command_center import (
-    DpmCommandCenterForwardRequest,
     DpmCommandCenterGatewayResponse,
-    DpmCommandCenterResolveExceptionRequest,
-    DpmExceptionSummaryGatewayResponse,
-    DpmExceptionSummaryRequest,
     DpmOutcomeReviewErrorDetail,
 )
 from app.middleware.correlation import correlation_id_var
@@ -81,161 +77,6 @@ async def get_command_center(
             "health_state": health_state,
             "limit": limit,
         },
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/monitoring/run-once",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="Run DPM mandate monitoring once",
-    description=(
-        "What: asks lotus-manage to evaluate a bounded set of refreshed mandate digital twins. "
-        "When: call this from an entitled Workbench command-center action or operator workflow. "
-        "How: Gateway forwards the request unchanged and returns manage's monitoring run state, "
-        "health results, exceptions, and lineage without discovering books or calculating health."
-    ),
-)
-async def run_monitoring_once(
-    request: DpmCommandCenterForwardRequest,
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().run_monitoring_once(
-        body=request.body,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/monitoring/runs",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="List DPM monitoring runs",
-    description=(
-        "What: lists manage-owned mandate monitoring runs newest first. When: use this for "
-        "command-center audit and operations drill-down. How: Gateway forwards search filters "
-        "to manage and preserves run status, source lineage, and supportability."
-    ),
-)
-async def list_monitoring_runs(
-    status_filter: str | None = Query(
-        default=None,
-        description="Optional terminal monitoring-run status filter.",
-        examples=["SUCCEEDED"],
-    ),
-    limit: int = Query(default=50, ge=1, le=200, description="Maximum runs to return."),
-    cursor: str | None = Query(default=None, description="Cursor from a previous page."),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().list_monitoring_runs(
-        filters={"status_filter": status_filter, "limit": limit, "cursor": cursor},
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/monitoring/runs/{monitoring_run_id}",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="Get DPM monitoring run",
-    description=(
-        "What: returns one manage-owned mandate monitoring run. When: use this for audit "
-        "drill-down from command-center latest-run or exception evidence. How: Gateway returns "
-        "the manage payload in a product envelope without changing health or exception truth."
-    ),
-)
-async def get_monitoring_run(
-    monitoring_run_id: str = Path(
-        ...,
-        description="Manage-owned mandate monitoring-run identifier.",
-        examples=["dmr_20260503_083000"],
-    ),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().get_monitoring_run(
-        monitoring_run_id=monitoring_run_id,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/exceptions",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="List DPM monitoring exceptions",
-    description=(
-        "What: lists manage-owned mandate monitoring exceptions. When: use this for Workbench "
-        "attention queues and operations triage by mandate, portfolio, or state. How: Gateway "
-        "preserves manage exception ids, severity, reason codes, state, and recommended action."
-    ),
-)
-async def list_monitoring_exceptions(
-    mandate_id: str | None = Query(default=None, description="Optional mandate id filter."),
-    portfolio_id: str | None = Query(default=None, description="Optional portfolio id filter."),
-    state: str | None = Query(
-        default=None,
-        description="Optional manage-published exception state filter.",
-        examples=["ACTIVE"],
-    ),
-    limit: int = Query(default=50, ge=1, le=200, description="Maximum exceptions to return."),
-    cursor: str | None = Query(default=None, description="Cursor from a previous page."),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().list_monitoring_exceptions(
-        filters={
-            "mandate_id": mandate_id,
-            "portfolio_id": portfolio_id,
-            "state": state,
-            "limit": limit,
-            "cursor": cursor,
-        },
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/exceptions/{exception_id}/resolve",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="Resolve DPM monitoring exception",
-    description=(
-        "What: forwards an exception resolution reason to lotus-manage. When: use this after a "
-        "PM, supervisor, or operator has reviewed the exception. How: Gateway does not close "
-        "exceptions locally; it returns the manage-owned resolved exception payload."
-    ),
-)
-async def resolve_monitoring_exception(
-    request: DpmCommandCenterResolveExceptionRequest,
-    exception_id: str = Path(
-        ...,
-        description="Manage-owned monitoring exception identifier.",
-        examples=["me_source_readiness_001"],
-    ),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().resolve_monitoring_exception(
-        exception_id=exception_id,
-        body={"resolution_reason": request.resolution_reason},
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/exceptions/{exception_id}/ai-summary",
-    response_model=DpmExceptionSummaryGatewayResponse,
-    summary="Request DPM exception AI summary",
-    description=(
-        "What: requests a governed lotus-ai exception-summary workflow-pack run from "
-        "manage-owned monitoring exception evidence. When: call this only for internal PM, "
-        "investment-control, or operations triage after the exception is visible in the command "
-        "center. How: Gateway reads the manage exception queue, builds a bounded no-raw-payload "
-        "evidence envelope for the selected exception, then executes lotus-ai "
-        "dpm_exception_summary.pack@v1 as lotus-gateway; Gateway does not generate narrative, "
-        "score PMs, approve trades, contact clients, route orders, or invent evidence."
-    ),
-)
-async def request_exception_summary(
-    request: DpmExceptionSummaryRequest,
-    exception_id: str = Path(
-        ...,
-        description="Manage-owned monitoring exception identifier for the bounded AI handoff.",
-        examples=["me_source_readiness_001"],
-    ),
-) -> DpmExceptionSummaryGatewayResponse:
-    return await dpm_command_center_service().request_exception_summary(
-        exception_id=exception_id,
-        request=request,
         correlation_id=correlation_id_var.get(),
     )
 
@@ -341,5 +182,3 @@ async def get_mandate_diff(
         filters={"from_version": from_version, "to_version": to_version},
         correlation_id=correlation_id_var.get(),
     )
-
-
